Create hospitalProject
This is the index.html of the HOSPITAL PROECT we are suppose to collaborate on. We are study the existing build and remove all components and frameworks as well as widgets, and do raw coding as replacement. The goal is to implement something close with the same look and feel of the original site with better functionality and an improved user experience.