Create CODE_OF_CONDUCT.md for our Hospital Project
I added a code of conduct to guide all contributors, and contributions as it’s a freelance project so, most people might become open to criticisms and unwanted advise from everyone involved in the project which might not be what you are used to therefore,  am adding this as a measure to ensure no one crosses the line, to infringe on your rights as volunteering coders! We are one and we are stronger together!